back/requirements/dev.txt
-r common.txt
django-cors-headers==3.0.1